+++ packages/components/src/molecules/Table/TableHead.tsx
@@ -1,5 +1,5 @@
-import type { HTMLAttributes } from 'react'
 import React, { forwardRef } from 'react'
+import type { HTMLAttributes, PropsWithChildren } from 'react'
 
 export interface TableHeadProps
   extends HTMLAttributes<HTMLTableSectionElement> {
@@ -7,23 +7,20 @@ export interface TableHeadProps
    * ID to find this component in testing tools (e.g.: cypress, testing library, and jest).
    */
   testId?: string
-  /**
-   * Children for TableHead components.
-   */
-  children: React.ReactNode
 }
 
-const TableHead = forwardRef<HTMLTableSectionElement, TableHeadProps>(
-  function TableHead(
-    { children, testId = 'fs-table-head', ...otherProps },
-    ref
-  ) {
-    return (
-      <thead ref={ref} data-testid={testId} data-fs-table-head {...otherProps}>
-        {children}
-      </thead>
-    )
-  }
-)
+const TableHead = forwardRef<
+  HTMLTableSectionElement,
+  PropsWithChildren<TableHeadProps>
+>(function TableHead(
+  { children, testId = 'fs-table-head', ...otherProps },
+  ref
+) {
+  return (
+    <thead ref={ref} data-testid={testId} data-fs-table-head {...otherProps}>
+      {children}
+    </thead>
+  )
+})
 
 export default TableHead