+++ CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.42-alpha.0](https://github.com/vtex/faststore/compare/v2.0.41-alpha.0...v2.0.42-alpha.0) (2023-01-10)
+
+
+### Chores
+
+* `Table` component's improvements ([#1592](https://github.com/vtex/faststore/issues/1592)) ([176a710](https://github.com/vtex/faststore/commit/176a710cf8d2917527cb4f99a8ddc956d841ae76))
+
+
+
+## [2.0.41-alpha.0](https://github.com/vtex/faststore/compare/v2.0.40-alpha.0...v2.0.41-alpha.0) (2023-01-10)
+
+
+### Features
+
+* Adds `Dropdown` component ([#1572](https://github.com/vtex/faststore/issues/1572)) ([edfe557](https://github.com/vtex/faststore/commit/edfe55797d51049b137c1d8c0d648f36d3ebf980))
+
+
+
+## [2.0.40-alpha.0](https://github.com/vtex/faststore/compare/v2.0.39-alpha.0...v2.0.40-alpha.0) (2023-01-10)
+
+
+### Documentation
+
+* Adds `Hero` component doc + Updates docs overview section ([#1598](https://github.com/vtex/faststore/issues/1598)) ([47b3199](https://github.com/vtex/faststore/commit/47b319904d0ff06fc5d3bacddbf5bfc319d451eb))
+
+
+
 ## [2.0.39-alpha.0](https://github.com/vtex/faststore/compare/v2.0.38-alpha.0...v2.0.39-alpha.0) (2023-01-09)